refactor: Remove scope_config from pagerduty as it is unused
