+++ backend/plugins/pagerduty/api/remote.go
@@ -131,10 +131,9 @@ func RemoteScopes(input *plugin.ApiResourceInput) (*plugin.ApiResourceOutput, er
 			Id:   service.Id,
 			Name: service.Name,
 			Data: models.Service{
-				Url:           service.HtmlUrl,
-				Id:            service.Id,
-				ScopeConfigId: 0, // this is not determined here
-				Name:          service.Name,
+				Url:  service.HtmlUrl,
+				Id:   service.Id,
+				Name: service.Name,
 			},
 		}
 		outputBody.Children = append(outputBody.Children, child)
