+++ backend/plugins/pagerduty/impl/impl.go
@@ -148,14 +148,6 @@ func (p PagerDuty) ApiResources() map[string]map[string]plugin.ApiResourceHandle
 			"PATCH":  api.UpdateScope,
 			"DELETE": api.DeleteScope,
 		},
-		"connections/:connectionId/scope-configs": {
-			"POST": api.CreateScopeConfig,
-			"GET":  api.GetScopeConfigList,
-		},
-		"connections/:connectionId/scope-configs/:id": {
-			"PATCH": api.UpdateScopeConfig,
-			"GET":   api.GetScopeConfig,
-		},
 	}
 }
